refactor: rename and reorganize package structure
