+++ linux/core.go
@@ -1,7 +1,7 @@
 //go:build linux
 // +build linux
 
-package main
+package linux
 
 import (
 	"bufio"
